refactor _filter.vue
