+++ src/composables/date.js
@@ -1,6 +1,6 @@
 import { getStorageKey } from '@/composables/runtime'
 
-const FILTER_STORAGE_KEY = getStorageKey('scm.filter.v1')
+export const FILTER_STORAGE_KEY = getStorageKey('scm.filter.v1')
 
 export function toLocalDate(rawDate) {
 
@@ -42,11 +42,11 @@ function stepToDate(step) {
       return date
 }
 
-function stepToISO(step) {
+export function stepToISO(step) {
       return formatToLayout(stepToDate(step))
 }
 
-function formatToLayout(date) {
+export function formatToLayout(date) {
       const year = date.getFullYear()
       const month = String(date.getMonth() + 1).padStart(2, '0')
       const day = String(date.getDate()).padStart(2, '0')
@@ -61,6 +61,17 @@ function formatToLayout(date) {
       const tzOffset = `${sign}${offsetHours}:${offsetMinutes}`
 
       return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}${tzOffset}`
+}
+
+export function formatToLayoutWithoutTimezone(date) {
+      const year = date.getFullYear()
+      const month = String(date.getMonth() + 1).padStart(2, '0')
+      const day = String(date.getDate()).padStart(2, '0')
+      const hours = String(date.getHours()).padStart(2, '0')
+      const minutes = String(date.getMinutes()).padStart(2, '0')
+      const seconds = String(date.getSeconds()).padStart(2, '0')
+
+      return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`
 }
